fix(geometry): bbox_to_mask position grid is exact at any image size
yy/xx were built via torch.arange(..., dtype=boxes.dtype) -- exact
pixel indices at whatever precision the caller's boxes happen to be.
float16 can only represent integers exactly up to 2048; past that,
consecutive positions collapse onto the same value, so a float16
`boxes` on an image taller or wider than 2048px silently mismasks rows
or columns near the collapse.

Reproduced: a box at [1000,1000]-[4090,3000] on a 4096x4096 image masks
6,185,091 pixels at float32 but a different count at float16, with an
entire extra row wrongly included at the y-boundary due to the rounding
collision.

Fix: build the position grid at a fixed float32 regardless of
boxes.dtype. float32-vs-boxes.dtype comparisons promote to float32
automatically, so no precision is lost anywhere else, and the returned
mask still casts back to boxes.dtype exactly as documented. float32/
float64 inputs are unaffected (already float32 already, or promotion
only goes up for float64).

Same bug class as the already-fixed rotary-embedding-torch#50 (integer
positions silently colliding under float16 past 2048) and the general
"position/index grid must be fixed-precision regardless of the
caller's own dtype" pattern.

Verified: float16 output now matches a float32 reference exactly
(byte-identical boolean mask) for a box spanning both axes past the
2048 collapse threshold; float32 input is byte-identical to before.
Added test_bbox_to_mask_matches_float32_reference_for_float16_boxes_on_a_large_image;
confirmed red on unpatched code (mismatched mask), green with the fix.
Full TestBbox2D suite: 21 passed, no regressions.

+++ kornia/geometry/bbox.py
@@ -335,8 +335,15 @@ def bbox_to_mask(boxes: torch.Tensor, width: int, height: int) -> torch.Tensor:
     # (`torch.any(...)` -> Python `if`) that blocked torch.compile fullgraph (e.g. RandomErasing,
     # which builds its mask through this function). Dropped; behaviour is byte-identical.
     # zero padding the surroundings
-    yy = torch.arange(height, device=boxes.device, dtype=boxes.dtype).view(height, 1)
-    xx = torch.arange(width, device=boxes.device, dtype=boxes.dtype).view(1, width)
+    # Built at a fixed float32 regardless of boxes.dtype: these are exact pixel
+    # indices, and float16 can only represent integers exactly up to 2048 --
+    # beyond that, consecutive positions collapse onto the same value, so a
+    # float16 `boxes` on an image taller/wider than 2048px silently mismasks
+    # rows/columns near the collapse. float32 vs boxes.dtype comparisons below
+    # promote to float32 automatically, so nothing downstream loses precision;
+    # the returned mask still casts back to boxes.dtype exactly as documented.
+    yy = torch.arange(height, device=boxes.device, dtype=torch.float32).view(height, 1)
+    xx = torch.arange(width, device=boxes.device, dtype=torch.float32).view(1, width)
     x_min = boxes[:, 0, 0].view(-1, 1, 1)
     y_min = boxes[:, 0, 1].view(-1, 1, 1)
     x_max = boxes[:, 2, 0].view(-1, 1, 1)
